requirements.txt
commando
fswrap
jinja2
pyparsing
pyyaml
boto